Allow switching between disc drive and FlippyDrive
Using L and R on the ‘game play’ screen, it’s now possible to switch between the disc drive and the FlippyDrive. There’s no visual feedback yet – that’s still yet to come.

To support this, all the `#if TEMP_TEST_DISC`s have been replaced with dynamic conditionals based on which device is active, with some simplifications where possible.