Add UI for switching between Flippy and disc drive
While on the cube’s ‘Game play’ menu (before pressing A to enter the menu), this displays a ‘L: Disc’ and ‘R: FlippyDrive’ UI, using icons to represent the two devices.

The selected device is highlighted, and this will update when a new one is selected (using L or R).

As not much gets drawn outside of the cube on this screen, the simplest way to do this was to draw it along with the button descriptions (making care to only show it while this screen is active). Additionally, this relies on the element alpha and blob systems, hence the added code for them.