Conditionally display ‘Force Copy’/‘Force Move’
If the save isn’t normally copyable or movable, the text in the Move/Copy/Erase menu will be replaced to indicate if we need to ‘Force Move’ or ‘Force Copy’ that save.

This is achieved by providing custom text data (in a STH0 blob) when drawing this menu. Unmodified text is copied verbatim, so it’s displayed in the current language.