Allow the language to be configured for NTSC IPLs
This adds a `ntsc_language` setting to the .ini, which can be set to `en`/`eng`/`english` or `ja`/`jpn`/`japanese` as preferred, or set to the new default of `auto` (setting the language according to the system’s region, as the IPL would do as standard).

This replaces the previous patch, which hardcodes NTSC IPLs to always be English.